services/auth-service/requirements.txt
flask==3.0.0
gunicorn==21.2.0
prometheus-client==0.19.0
python-json-logger==2.0.7
PyJWT==2.8.0
bcrypt==4.1.2
